// services/user-remote/src/generated/types/users_pb.d.ts
export class UserModel extends jspb.Message { 

    hasId(): boolean;
    clearId(): void;
    getId(): google_protobuf_wrappers_pb.StringValue | undefined;
    setId(value?: google_protobuf_wrappers_pb.StringValue): UserModel;

    hasUsername(): boolean;
    clearUsername(): void;
    getUsername(): google_protobuf_wrappers_pb.StringValue | undefined;
    setUsername(value?: google_protobuf_wrappers_pb.StringValue): UserModel;

    hasEmail(): boolean;
    clearEmail(): void;
    getEmail(): google_protobuf_wrappers_pb.StringValue | undefined;
    setEmail(value?: google_protobuf_wrappers_pb.StringValue): UserModel;

    hasPassword(): boolean;
    clearPassword(): void;
    getPassword(): google_protobuf_wrappers_pb.StringValue | undefined;
    setPassword(value?: google_protobuf_wrappers_pb.StringValue): UserModel;

    hasPicture(): boolean;
    clearPicture(): void;
    getPicture(): google_protobuf_wrappers_pb.StringValue | undefined;
    setPicture(value?: google_protobuf_wrappers_pb.StringValue): UserModel;

    hasToken(): boolean;
    clearToken(): void;
    getToken(): google_protobuf_wrappers_pb.StringValue | undefined;
    setToken(value?: google_protobuf_wrappers_pb.StringValue): UserModel;

    hasUserwallet(): boolean;
    clearUserwallet(): void;
    getUserwallet(): UserWallet | undefined;
    setUserwallet(value?: UserWallet): UserModel;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): UserModel.AsObject;
    static toObject(includeInstance: boolean, msg: UserModel): UserModel.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: UserModel, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): UserModel;
    static deserializeBinaryFromReader(message: UserModel, reader: jspb.BinaryReader): UserModel;
}

export namespace UserModel {
    export type AsObject = {
        Id?: google_protobuf_wrappers_pb.StringValue.AsObject,
        username?: google_protobuf_wrappers_pb.StringValue.AsObject,
        email?: google_protobuf_wrappers_pb.StringValue.AsObject,
        password?: google_protobuf_wrappers_pb.StringValue.AsObject,
        picture?: google_protobuf_wrappers_pb.StringValue.AsObject,
        token?: google_protobuf_wrappers_pb.StringValue.AsObject,
        userwallet?: UserWallet.AsObject,
    }
}
// ...
export class UserWallet extends jspb.Message { 

    hasToken(): boolean;
    clearToken(): void;
    getToken(): google_protobuf_wrappers_pb.StringValue | undefined;
    setToken(value?: google_protobuf_wrappers_pb.StringValue): UserWallet;

    hasAmount(): boolean;
    clearAmount(): void;
    getAmount(): google_protobuf_wrappers_pb.StringValue | undefined;
    setAmount(value?: google_protobuf_wrappers_pb.StringValue): UserWallet;

    hasTousd(): boolean;
    clearTousd(): void;
    getTousd(): google_protobuf_wrappers_pb.StringValue | undefined;
    setTousd(value?: google_protobuf_wrappers_pb.StringValue): UserWallet;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): UserWallet.AsObject;
    static toObject(includeInstance: boolean, msg: UserWallet): UserWallet.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: UserWallet, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): UserWallet;
    static deserializeBinaryFromReader(message: UserWallet, reader: jspb.BinaryReader): UserWallet;
}

export namespace UserWallet {
    export type AsObject = {
        token?: google_protobuf_wrappers_pb.StringValue.AsObject,
        amount?: google_protobuf_wrappers_pb.StringValue.AsObject,
        tousd?: google_protobuf_wrappers_pb.StringValue.AsObject,
    }
}

export class GetUserByIdForm extends jspb.Message { 

    hasId(): boolean;
    clearId(): void;
    getId(): google_protobuf_wrappers_pb.StringValue | undefined;
    setId(value?: google_protobuf_wrappers_pb.StringValue): GetUserByIdForm;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): GetUserByIdForm.AsObject;
    static toObject(includeInstance: boolean, msg: GetUserByIdForm): GetUserByIdForm.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: GetUserByIdForm, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): GetUserByIdForm;
    static deserializeBinaryFromReader(message: GetUserByIdForm, reader: jspb.BinaryReader): GetUserByIdForm;
}

export namespace GetUserByIdForm {
    export type AsObject = {
        Id?: google_protobuf_wrappers_pb.StringValue.AsObject,
    }
}

export class GetAllUsersRequest extends jspb.Message { 

    hasLimit(): boolean;
    clearLimit(): void;
    getLimit(): google_protobuf_wrappers_pb.Int32Value | undefined;
    setLimit(value?: google_protobuf_wrappers_pb.Int32Value): GetAllUsersRequest;

    hasPage(): boolean;
    clearPage(): void;
    getPage(): google_protobuf_wrappers_pb.Int32Value | undefined;
    setPage(value?: google_protobuf_wrappers_pb.Int32Value): GetAllUsersRequest;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): GetAllUsersRequest.AsObject;
    static toObject(includeInstance: boolean, msg: GetAllUsersRequest): GetAllUsersRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: GetAllUsersRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): GetAllUsersRequest;
    static deserializeBinaryFromReader(message: GetAllUsersRequest, reader: jspb.BinaryReader): GetAllUsersRequest;
}

export namespace GetAllUsersRequest {
    export type AsObject = {
        limit?: google_protobuf_wrappers_pb.Int32Value.AsObject,
        page?: google_protobuf_wrappers_pb.Int32Value.AsObject,
    }
}

export class GetUsersRequest extends jspb.Message { 

    hasLimit(): boolean;
    clearLimit(): void;
    getLimit(): google_protobuf_wrappers_pb.Int32Value | undefined;
    setLimit(value?: google_protobuf_wrappers_pb.Int32Value): GetUsersRequest;

    hasPage(): boolean;
    clearPage(): void;
    getPage(): google_protobuf_wrappers_pb.Int32Value | undefined;
    setPage(value?: google_protobuf_wrappers_pb.Int32Value): GetUsersRequest;

    hasFilter(): boolean;
    clearFilter(): void;
    getFilter(): google_protobuf_wrappers_pb.StringValue | undefined;
    setFilter(value?: google_protobuf_wrappers_pb.StringValue): GetUsersRequest;

    hasMatch(): boolean;
    clearMatch(): void;
    getMatch(): google_protobuf_wrappers_pb.StringValue | undefined;
    setMatch(value?: google_protobuf_wrappers_pb.StringValue): GetUsersRequest;

    hasIdlist(): boolean;
    clearIdlist(): void;
    getIdlist(): GetUserByIdForm | undefined;
    setIdlist(value?: GetUserByIdForm): GetUsersRequest;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): GetUsersRequest.AsObject;
    static toObject(includeInstance: boolean, msg: GetUsersRequest): GetUsersRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: GetUsersRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): GetUsersRequest;
    static deserializeBinaryFromReader(message: GetUsersRequest, reader: jspb.BinaryReader): GetUsersRequest;
}

export namespace GetUsersRequest {
    export type AsObject = {
        limit?: google_protobuf_wrappers_pb.Int32Value.AsObject,
        page?: google_protobuf_wrappers_pb.Int32Value.AsObject,
        filter?: google_protobuf_wrappers_pb.StringValue.AsObject,
        match?: google_protobuf_wrappers_pb.StringValue.AsObject,
        Idlist?: GetUserByIdForm.AsObject,
    }
}

export class GetUserByIdRequest extends jspb.Message { 

    hasId(): boolean;
    clearId(): void;
    getId(): google_protobuf_wrappers_pb.StringValue | undefined;
    setId(value?: google_protobuf_wrappers_pb.StringValue): GetUserByIdRequest;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): GetUserByIdRequest.AsObject;
    static toObject(includeInstance: boolean, msg: GetUserByIdRequest): GetUserByIdRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: GetUserByIdRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): GetUserByIdRequest;
    static deserializeBinaryFromReader(message: GetUserByIdRequest, reader: jspb.BinaryReader): GetUserByIdRequest;
}

export namespace GetUserByIdRequest {
    export type AsObject = {
        id?: google_protobuf_wrappers_pb.StringValue.AsObject,
    }
}
// ...
export class FollowUserRequest extends jspb.Message { 

    hasId(): boolean;
    clearId(): void;
    getId(): google_protobuf_wrappers_pb.StringValue | undefined;
    setId(value?: google_protobuf_wrappers_pb.StringValue): FollowUserRequest;

    hasCurrentUserId(): boolean;
    clearCurrentUserId(): void;
    getCurrentUserId(): google_protobuf_wrappers_pb.StringValue | undefined;
    setCurrentUserId(value?: google_protobuf_wrappers_pb.StringValue): FollowUserRequest;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): FollowUserRequest.AsObject;
    static toObject(includeInstance: boolean, msg: FollowUserRequest): FollowUserRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: FollowUserRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): FollowUserRequest;
    static deserializeBinaryFromReader(message: FollowUserRequest, reader: jspb.BinaryReader): FollowUserRequest;
}

export namespace FollowUserRequest {
    export type AsObject = {
        id?: google_protobuf_wrappers_pb.StringValue.AsObject,
        currentUserId?: google_protobuf_wrappers_pb.StringValue.AsObject,
    }
}

export class UnfollowUserRequest extends jspb.Message { 

    hasId(): boolean;
    clearId(): void;
    getId(): google_protobuf_wrappers_pb.StringValue | undefined;
    setId(value?: google_protobuf_wrappers_pb.StringValue): UnfollowUserRequest;

    hasCurrentUserId(): boolean;
    clearCurrentUserId(): void;
    getCurrentUserId(): google_protobuf_wrappers_pb.StringValue | undefined;
    setCurrentUserId(value?: google_protobuf_wrappers_pb.StringValue): UnfollowUserRequest;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): UnfollowUserRequest.AsObject;
    static toObject(includeInstance: boolean, msg: UnfollowUserRequest): UnfollowUserRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: UnfollowUserRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): UnfollowUserRequest;
    static deserializeBinaryFromReader(message: UnfollowUserRequest, reader: jspb.BinaryReader): UnfollowUserRequest;
}

export namespace UnfollowUserRequest {
    export type AsObject = {
        id?: google_protobuf_wrappers_pb.StringValue.AsObject,
        currentUserId?: google_protobuf_wrappers_pb.StringValue.AsObject,
    }
}